requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
gunicorn==21.2.0
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.21.0
openai==0.28.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyJWT==2.8.0
pyngrok==7.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
srt==3.5.3
starlette==0.27.0
tqdm==4.66.1
twilio==8.10.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
vosk==0.3.44
websockets==12.0
yarl==1.9.2
